# net/netip: inlinable AsSlice