benchmark                    old ns/op     new ns/op     delta
BenchmarkAsSliceAddrv4-4     19.6          2.95          -84.95%
BenchmarkAsSliceAddrv6-4     57.4          3.13          -94.55%

benchmark                    old allocs     new allocs     delta
BenchmarkAsSliceAddrv4-4     1              0              -100.00%
BenchmarkAsSliceAddrv6-4     1              0              -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkAsSliceAddrv4-4     4             0             -100.00%
BenchmarkAsSliceAddrv6-4     16            0             -100.00%